# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user and return auth token",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "User Name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "description": "User Email",
                                        "type": "string",
                                        "example": "john@mail.com"
                                    },
                                    "password": {
                                        "description": "User Password",
                                        "type": "string",
                                        "example": "super-secret"
                                    },
                                    "password_confirmation": {
                                        "description": "User Password Confirmation",
                                        "type": "string",
                                        "example": "super-secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "plainTextToken": {
                                            "type": "string",
                                            "example": "1|wLn8vrVU5g8N9u13WdC7qcT5KZk6yi7jo8lcqC5wfd07b1f2"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email has already been taken."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "The email has already been taken."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log In user and return auth token",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User Email",
                                        "type": "string",
                                        "example": "john@mail.com"
                                    },
                                    "password": {
                                        "description": "User Password",
                                        "type": "string",
                                        "example": "super-secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful log in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "plainTextToken": {
                                            "type": "string",
                                            "example": "1|wLn8vrVU5g8N9u13WdC7qcT5KZk6yi7jo8lcqC5wfd07b1f2"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The provided credentials are incorrect"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": [
                                                    "The provided credentials are incorrect"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout the authenticated user",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "204": {
                        "description": "Successfully logged out. No content returned."
                    },
                    "401": {
                        "description": "Unauthorized – invalid or missing token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}